requirements.txt
python-dotenv==1.0.0
requests==2.31.0
flask==3.0.0
gunicorn==21.2.0
google-generativeai==0.3.0
huggingface-hub==0.20.1
aiohttp==3.9.1
asyncio-contextmanager==1.0.0
